Rename tests, extract test vectors
