+++ packages/util-crypto/src/nacl/tweetnacl-secretbox-data.spec.ts
@@ -1,25 +1,7 @@
 // Copyright 2017-2023 @polkadot/util-crypto authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-/// <reference types="@polkadot/dev-test/globals.d.ts" />
-
-// Adapted from https://github.com/dchest/tweetnacl-js/blob/6a9594a35a27f9c723c5f1c107e376d1c65c23b3/test/04-secretbox.quick.js
-// Adapted from https://github.com/dchest/tweetnacl-js/blob/6a9594a35a27f9c723c5f1c107e376d1c65c23b3/test/04-secretbox.js
-//
-// Original headers:
-//
-// Ported in 2014 by Dmitry Chestnykh and Devi Mandiri.
-// Public domain.
-//
-// Implementation derived from TweetNaCl version 20140427.
-// See for details: http://tweetnacl.cr.yp.to/
-
-import { stringToU8a, u8aEq } from '@polkadot/util';
-
-import { base64Decode } from '../base64/index.js';
-import { naclSecretbox, naclSecretboxOpen } from './tweetnacl.js';
-
-const TEST_DATA = [
+export const TEST_DATA = [
   [
     'givKPH4F/eDcIEUZcws1+BIWqcnx35Ul4qkA7Ilxj1c=',
     'crkCCNKADjatFscwlBoDjXw62dhwMNMp',
@@ -4629,139 +4611,3 @@ const TEST_DATA = [
     '4GRhghuLJqaA3lxM2G4nPEOZuEzirdZUk9+4J2mcx9OGKZA84JBDmomUpST0jCTVv2kWdiAv/TGpJiFWpdojB/85g3T5XUogNiFgWV/6d+UkPK1JATqKAKBtUNYBQp43Cg5HdCNDJXHLBmyY25IyseUzlah9znfUQg/5ARvLCHrwG8FrI+8j7ZZQN4iPaEWwA0VGXzQbaJgm4JIQQ8eCSbMFJJ8Oa19iaBKlPXHCIvZOiru32KUFZh0jOy9O8SwU+zyFVaeF9wpbb9SV0j4dV4MDRPbqHbb+bj2XNE7tKWPp716PIPobFuB1+Z4OsM107AKfrLliALBZZKrb63/IxFT9RYC/6Xei4nfHhhqv+4MLcslD7nPr1NtLOG7iyCNAQ+4vRCHzupmwguazPTOrxBuUYBePLzF44yVmCU94SM1bXvdRwf6XvpedEF0Q7LUEhVaBCQYrk+zgCcg4b9+IjZSF1KLi7K/aOwZfOfXKzleNI66pb2+Ey0StVhw0U9gWUrDwGn8XChF/8c99csBkejUSV+voV0zVtuR2wzCjR5ZnaNvIZipWf/l5Iflg/egsHlOCWTAWwMyt1HUKIAk3pxljBE7tqKSax6buECmsh6d59tF5vd2+2wp8zER3wfSEfLj7lRTQ9FUc/4z8dZPOP6v7qKX9LwMD64kM7iWGbJMUzd7L42+HNoBJgVyMvbq4MDV8iesXaGUab2Ad6Sfx9fzRUn/HtAsvKx3N/fHpUOnoZT5A/y3xkbUmLxFFqp6yih5LpZoXK4qf6GoYlxZbuk+krkMD4UrsxH8of2gHHil6OQD0tOIkilMIHPeJFzGrlTvX5wl4sricArnDG/qEo9Bu7CR1KF7b9UgWcR5rUhPo9O6PR+OhnjDiPlOjBjoQ2HKvIoLWsCMz8FkgYyNyosmW7OLJOaS8dF+IV8cRSBWQjToYtT6nB3KIKWEtAwtFvYJxFOtOQwNKtlmgLZGMDjEttdwieW+t6ZRj1fZ5tmybu2bBIzJ74obL3lUaSyldnxBE116spRKw6rlgmT0U'
   ]
 ];
-
-describe('tweetnacl', (): void => {
-  describe('secretbox.quick from tweetnacl', (): void => {
-    it('naclSecretbox and naclSecretbox.open', (): void => {
-      const key = new Uint8Array(32);
-      const nonce = new Uint8Array(24);
-
-      for (let i = 0; i < key.length; i++) {
-        key[i] = i & 0xff;
-      }
-
-      for (let i = 0; i < nonce.length; i++) {
-        nonce[i] = (32 + i) & 0xff;
-      }
-
-      const msg = stringToU8a('message to encrypt');
-      const box = naclSecretbox(msg, nonce, key);
-      const openedMsg = naclSecretboxOpen(box, nonce, key);
-
-      expect(
-        !!openedMsg && u8aEq(openedMsg, msg)
-      ).toEqual(true);
-    });
-
-    it('naclSecretbox.open with invalid box', (): void => {
-      const key = new Uint8Array(32);
-      const nonce = new Uint8Array(24);
-
-      expect(
-        naclSecretboxOpen(new Uint8Array(0), nonce, key)
-      ).toBe(null);
-
-      expect(
-        naclSecretboxOpen(new Uint8Array(10), nonce, key)
-      ).toBe(null);
-
-      expect(
-        naclSecretboxOpen(new Uint8Array(100), nonce, key)
-      ).toBe(null);
-    });
-
-    it('naclSecretbox.open with invalid nonce', (): void => {
-      const key = new Uint8Array(32);
-      const nonce = new Uint8Array(24);
-
-      for (let i = 0; i < nonce.length; i++) {
-        nonce[i] = i & 0xff;
-      }
-
-      const msg = stringToU8a('message to encrypt');
-      const box = naclSecretbox(msg, nonce, key);
-      const unbox = naclSecretboxOpen(box, nonce, key);
-
-      expect(
-        !!unbox && u8aEq(unbox, msg)
-      ).toBe(true);
-
-      nonce[0] = 255;
-
-      expect(
-        naclSecretboxOpen(box, nonce, key)
-      ).toBe(null);
-    });
-
-    it('naclSecretbox.open with invalid key', (): void => {
-      const key = new Uint8Array(32);
-
-      for (let i = 0; i < key.length; i++) {
-        key[i] = i & 0xff;
-      }
-
-      const nonce = new Uint8Array(24);
-      const msg = stringToU8a('message to encrypt');
-      const box = naclSecretbox(msg, nonce, key);
-      const unbox = naclSecretboxOpen(box, nonce, key);
-
-      expect(
-        !!unbox && u8aEq(unbox, msg)
-      ).toBe(true);
-
-      key[0] = 255;
-
-      expect(
-        naclSecretboxOpen(box, nonce, key)
-      ).toBe(null);
-    });
-
-    it('naclSecretbox with message lengths of 0 to 1024', (): void => {
-      const key = new Uint8Array(32);
-      const nonce = new Uint8Array(24);
-      const fullMsg = new Uint8Array(1024);
-
-      for (let i = 0; i < key.length; i++) {
-        key[i] = i & 0xff;
-      }
-
-      for (let i = 0; i < fullMsg.length; i++) {
-        fullMsg[i] = i & 0xff;
-      }
-
-      for (let i = 0; i < fullMsg.length; i++) {
-        const msg = fullMsg.subarray(0, i);
-        const box = naclSecretbox(msg, nonce, key);
-        const unbox = naclSecretboxOpen(box, nonce, key);
-
-        expect(
-          !!unbox && u8aEq(unbox, msg)
-        ).toBe(true);
-      }
-    });
-  });
-
-  describe('secretbox from tweetnacl', (): void => {
-    for (let i = 0; i < TEST_DATA.length; i++) {
-      const [keyBase64, nonceBase64, msgBase64, expBase64] = TEST_DATA[i];
-
-      it(`is ok on "${msgBase64}"`, (): void => {
-        const key = base64Decode(keyBase64);
-        const nonce = base64Decode(nonceBase64);
-        const msg = base64Decode(msgBase64);
-        const goodBox = base64Decode(expBase64);
-        const box = naclSecretbox(msg, nonce, key);
-
-        expect(
-          u8aEq(box, goodBox)
-        ).toBe(true);
-
-        const unbox = naclSecretboxOpen(goodBox, nonce, key);
-
-        expect(
-          !!unbox && u8aEq(unbox, msg)
-        ).toBe(true);
-      });
-    }
-  });
-});
